.history/requirements_20250428124034.txt
python-telegram-bot==20.7
telethon==1.33.1
aiohttp==3.9.1
watchdog==3.0.0
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
pyTelegramBotAPI-4.26.0